+++ packages/assets-controllers/src/AssetsContractController.ts
@@ -18,6 +18,7 @@ import { getKnownPropertyNames, type Hex } from '@metamask/utils';
 import type BN from 'bn.js';
 
 import {
+  STAKING_CONTRACT_ADDRESS_BY_CHAINID,
   SupportedStakedBalanceNetworks,
   SupportedTokenDetectionNetworks,
 } from './assetsUtil';
@@ -74,13 +75,6 @@ export const SINGLE_CALL_BALANCES_ADDRESS_BY_CHAINID = {
   [SupportedTokenDetectionNetworks.moonriver]:
     '0x6aa75276052d96696134252587894ef5ffa520af',
 } as const satisfies Record<Hex, string>;
-
-export const STAKING_CONTRACT_ADDRESS_BY_CHAINID = {
-  [SupportedStakedBalanceNetworks.mainnet]:
-    '0x4fef9d741011476750a243ac70b9789a63dd47df',
-  [SupportedStakedBalanceNetworks.hoodi]:
-    '0xe96ac18cfe5a7af8fe1fe7bc37ff110d88bc67ff',
-} as Record<Hex, string>;
 
 export const MISSING_PROVIDER_ERROR =
   'AssetsContractController failed to set the provider correctly. A provider must be set for this method to be available';
