+++ packages/assets-controllers/src/balances.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { AccountWalletType, AccountGroupType } from '@metamask/account-api';
 
-import { STAKING_CONTRACT_ADDRESS_BY_CHAINID } from './AssetsContractController';
+import { STAKING_CONTRACT_ADDRESS_BY_CHAINID } from './assetsUtil';
 import {
   calculateBalanceForAllWallets,
   calculateBalanceChangeForAllWallets,
